order dao done somhow

+++ include/dao/RestaurantDAO.hpp
@@ -3,6 +3,8 @@
 
 #include "Restaurant.hpp"
 #include <sqlite3.h>
+#include <vector>
+#include <memory>
 
 class RestaurantDAO
 {
@@ -11,21 +13,21 @@ private:
     sqlite3 *db;
 
 public:
-    explicit RestaurantDAO(sqlite3 *); // using expilicit won't let compiler automatically change a sqlite3* to a restaurantdao.
+    explicit RestaurantDAO(sqlite3 *); // using explicit won't let compiler automatically convert sqlite3* to RestaurantDAO.
 
     // create
     void CreateTable();
     bool Create(const Restaurant &);
 
     // read
-    std::unique_ptr<Restaurant> ReadById(long); // we use unique ptr for handeling the ownership and stop unwanted copying handeling the memory(delete)
+    std::unique_ptr<Restaurant> ReadById(long long); // unique_ptr for handling ownership and preventing unwanted copying
     std::vector<std::unique_ptr<Restaurant>> ReadAll();
 
     // update
     bool Update(const Restaurant &);
 
     // delete
-    bool Delete(long);
+    bool Delete(long long);
 };
 
 #endif
